"forego" -> "eschew"
"Forgo" and "eschew" are synonyms but there's a subtle difference: "forgo" implies that we're giving up something — often reluctantly; while "eschew" means just that we're intentionally not using it, without implication of reluctance or hardship.

(Aside: I would consider "forego" to be, in this case, just an alternate spelling of "forgo". Apparently someone says this is "incorrect", but they also say that Charles Dickens and Emily Brontë agree with me. And that Brontë's use of it in the "giving up something" sense in Wuthering Heights is a "misspelling". I mean come on. Also Oxford Languages agrees that it's an alternate spelling. So that *isn't* the reason we're changing it.)